wip: Added Dummy Data for Train A to work with
After discussing with both Train A and B about how
data is to be parsed through to the front end, our
train constructed a possible type structure to be
used. Using this type structure, and some changes
afterwards, and the use of some ChatGPT, some fake
user data has been generated to be used for some
initial testing with apache echarts.
It is intended to be used temporarily.